+++ signalwire/relay/calling/results/disconnect_result.py
@@ -1,0 +1,5 @@
+from . import BaseResult
+
+class DisconnectResult(BaseResult):
+  def __init__(self, component):
+    super().__init__(component)